Enhance interaction event handling and validation
- Introduced a new `validate_interaction_event_type_filter` function to ensure valid interaction event types, improving user input validation.
- Updated the `list_interaction_events` method to utilize the new validation function, enhancing robustness in event querying.
- Added comprehensive tests for the validation function to ensure correct handling of valid and invalid event types, improving overall test coverage.
- Refactored SQL schema to add new columns for interaction sessions and turns, enhancing data management capabilities.
