+++ bitloops/src/graphql/types/repository.rs
@@ -3,6 +3,7 @@ use async_graphql::{ComplexObject, Context, ID, Result, SimpleObject};
 use crate::graphql::{
     DevqlGraphqlContext, ResolverScope, backend_error, bad_cursor_error, bad_user_input_error,
 };
+use crate::host::interactions::types::InteractionEventType;
 
 use super::interaction::{InteractionEventObject, InteractionSessionObject, InteractionTurnObject};
 use super::{
@@ -423,6 +424,7 @@ impl Repository {
         since: Option<String>,
         first: Option<i32>,
     ) -> Result<Vec<InteractionEventObject>> {
+        validate_interaction_event_type_filter(event_type.as_deref())?;
         ctx.data_unchecked::<DevqlGraphqlContext>()
             .list_interaction_events(
                 &self.scope,
@@ -434,5 +436,38 @@ impl Repository {
             )
             .await
             .map_err(|err| backend_error(format!("failed to query interaction events: {err:#}")))
+    }
+}
+
+fn validate_interaction_event_type_filter(event_type: Option<&str>) -> Result<()> {
+    if let Some(value) = event_type.map(str::trim).filter(|value| !value.is_empty())
+        && InteractionEventType::parse(value).is_none()
+    {
+        return Err(bad_user_input_error(format!(
+            "invalid interaction event type `{value}`"
+        )));
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn invalid_interaction_event_type_filter_is_bad_user_input() {
+        let err = validate_interaction_event_type_filter(Some("not_real")).unwrap_err();
+        assert!(
+            format!("{err:?}").contains("BAD_USER_INPUT"),
+            "expected BAD_USER_INPUT, got: {err:?}"
+        );
+        assert_eq!(err.message, "invalid interaction event type `not_real`");
+    }
+
+    #[test]
+    fn known_interaction_event_type_filter_is_allowed() {
+        validate_interaction_event_type_filter(Some("turn_end")).expect("valid filter");
+        validate_interaction_event_type_filter(Some("  ")).expect("blank filter");
+        validate_interaction_event_type_filter(None).expect("missing filter");
     }
 }
